// backend/prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const MessageRole = {
  User: 'User',
  Assistant: 'Assistant'
} as const

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole]


export const AuthProvider = {
  google: 'google',
  github: 'github'
} as const

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider]